add navigator share for video clips
